+++ apps/electron/src/window/overlay-manager.ts
@@ -1,6 +1,7 @@
 import { Store } from '../store';
 import { BrowserWindow, BrowserWindowConstructorOptions } from 'electron';
 import { Constants } from '../constants';
+import { join } from 'path';
 
 export class OverlayManager {
 
@@ -28,9 +29,8 @@ export class OverlayManager {
       width: dimensions.x,
       height: dimensions.y,
       webPreferences: {
-        nodeIntegration: true,
         backgroundThrottling: false,
-        contextIsolation: false
+        preload: join(__dirname, '../preload.js')
       }
     };
     Object.assign(opts, this.store.get(`overlay:${url}:bounds`, {}));
@@ -42,11 +42,6 @@ export class OverlayManager {
 
     overlay.once('ready-to-show', () => {
       overlay.show();
-    });
-
-    // save window size and position
-    overlay.on('close', () => {
-      this.afterOverlayClose(url);
     });
 
 
@@ -75,9 +70,18 @@ export class OverlayManager {
   }
 
   sendToOverlay(uri: string, channel: string, payload: any): void {
-    if (this.openedOverlays[uri] !== undefined) {
+    if (this.openedOverlays[uri] !== undefined && !this.openedOverlays[uri].isDestroyed()) {
       this.openedOverlays[uri].webContents.send(channel, payload);
     }
+  }
+
+  closeOverlay(url: string): void {
+    const overlay = this.openedOverlays[url];
+    if (!overlay) {
+      return;
+    }
+    overlay.close();
+    this.afterOverlayClose(url);
   }
 
   persistOverlays(): void {